created header in a different project and copy, pasted it into this one
Wrote the HTML and CSS code for the header. It is not currenlty interactive but I will figure it out later as I go. learned to create <div> tags and used them to separate the different sections of the header. Used <a> and <li> inside a <ul> to create the navigation. I will be adding links to those when I figure out how to do it.